refactor(ui): move API feedback logic to local component directory
Moved `feedbackApi.ts` from `apps/site-2026/src/api/` to `apps/site-2026/src/components/Feedback/` to improve component encapsulation. The `FeedbackPayload` structure has been updated, the `form_version` field has been added and imports have been corrected.
